[fix] クリーンアーキテクチャっぽくファイル分け
