+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Uniiva">
-        <activity android:name=".main.MainActivity">
+        <activity android:name="net.uniiva.view.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
