+++ app/src/main/java/net/uniiva/domain/main/fragment/home/HomeDomainInterface.kt
@@ -1,0 +1,8 @@
+package net.uniiva.domain.main.fragment.home
+
+interface HomeDomainInterface {
+
+    fun signIn(): String
+
+    fun signOut()
+}
