+++ app/src/main/java/net/uniiva/presenter/main/fragment/home/HomePresenterInterface.kt
@@ -1,0 +1,6 @@
+package net.uniiva.presenter.main.fragment.home
+
+interface HomePresenterInterface {
+
+    fun setView()
+}
