+++ app/src/main/java/net/uniiva/view/main/fragment/CircleFragment.kt
@@ -1,15 +1,15 @@
-package net.uniiva.main.mypage
+package net.uniiva.view.main.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import net.uniiva.databinding.FragmentMypageBinding
+import net.uniiva.databinding.FragmentCircleBinding
 
-class MyPageFragment : Fragment() {
+class CircleFragment : Fragment() {
 
-    private var _binding: FragmentMypageBinding? = null
+    private var _binding: FragmentCircleBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -17,7 +17,7 @@ class MyPageFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentMypageBinding.inflate(inflater, container, false)
+        _binding = FragmentCircleBinding.inflate(inflater, container, false)
         return binding.root
     }
 }
