+++ app/src/main/java/net/uniiva/view/main/fragment/HomeFragment.kt
@@ -1,0 +1,37 @@
+package net.uniiva.view.main.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import net.uniiva.databinding.FragmentHomeBinding
+import net.uniiva.presenter.main.fragment.home.HomePresenter
+
+class HomeFragment : Fragment() {
+
+    private var _binding: FragmentHomeBinding? = null
+    private val binding get() = _binding!!
+
+    private lateinit var homePresenter: HomePresenter
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        homePresenter = HomePresenter(binding, requireActivity() as AppCompatActivity)
+
+        homePresenter.setView()
+
+    }
+}
