+++ app/src/main/res/navigation/main_bottom_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="net.uniiva.main.home.HomeFragment"
+        android:name="net.uniiva.view.main.fragment.HomeFragment"
         android:label="home_fragment"
         tools:layout="@layout/fragment_home"/>
 
     <fragment
         android:id="@+id/navigation_mypage"
-        android:name="net.uniiva.main.mypage.MyPageFragment"
+        android:name="net.uniiva.view.main.fragment.MyPageFragment"
         android:label="navigation_mypage"
         tools:layout="@layout/fragment_mypage"/>
 
     <fragment
         android:id="@+id/navigation_circle"
-        android:name="net.uniiva.main.circle.CircleFragment"
+        android:name="net.uniiva.view.main.fragment.CircleFragment"
         android:label="circle_fragment"
         tools:layout="@layout/fragment_circle"/>
 </navigation>
